client-sdks/python/test_requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
requests-mock>=1.10.0
coverage>=7.0.0
